refactor(hooks): extract circuit title logic into useCircuitTitle hook
The circuit title logic was previously embedded directly in the RunFrameWithApi component. This commit extracts it into a dedicated useCircuitTitle hook to improve code modularity and reusability.
